A notebook to show how SLAM works step by step
FEATURE: A notebook to show how we can track the motion of a robot and localize the landmarks using a SLAM algorithm.

We construct a constraint matrix using the sensing information to solve for the landmark position and the robot's motion. 
The robot's class is defined in robot.py, and the motion data (to mimic the sensor measurement) is generated using functions in helper.py.